backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic[email]==2.6.3
pydantic-settings==2.2.1
sqlalchemy==2.0.28
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.3
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
loguru==0.7.2
boto3>=1.34.100
openai>=1.30.0

gunicorn==21.2.0
pytest==8.0.2
pytest-asyncio==0.23.5
black==24.2.0
isort==5.13.2
mypy==1.8.0
ruff==0.3.0
